Only do the download clear when not building with debian snapshots
git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@5548 b52b6941-3400-464c-9f42-43200397181c

+++ rebeccablackos_files/tmp/configure_phase1.sh
@@ -195,7 +195,7 @@ fi
 
 #Remove packages that can no longer be downloaded to preserve space
 ESSENTIALOBSOLETEPACKAGECOUNT=$(aptitude search '~o~E' |wc -l)
-if [[ $ESSENTIALOBSOLETEPACKAGECOUNT == 0 ]]
+if [[ $ESSENTIALOBSOLETEPACKAGECOUNT == 0 && ! -e /tmp/buildcore_revisions.txt ]]
 then
   apt-get autoclean -o APT::Clean-Installed=off                          2>&1 |tee -a "$PACKAGEOPERATIONLOGDIR"/Downloads/purge_obsolete.log
 else
